requirements.txt
langchain
pinecone-client
tiktoken
openai
fastapi
langchain-community
langchain-openai
uvicorn
langchain_core
langchain_pinecone
uvicorn[standard]